requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
pillow==11.1.0
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0